+++ pyproject.toml
@@ -17,13 +17,13 @@ psycopg = ["psycopg>=3.2.0"]
 
 [project.entry-points."sqlalchemy.dialects"]
 # The default dialect uses psycopg (psycopg3) 
-auroradsql = "auroradsql_sqlalchemy.psycopg:AuroraDSQLDialect_psycopg"
-"auroradsql.psycopg2" = "auroradsql_sqlalchemy.psycopg2:AuroraDSQLDialect_psycopg2"
+auroradsql = "aurora_dsql_sqlalchemy.psycopg:AuroraDSQLDialect_psycopg"
+"auroradsql.psycopg2" = "aurora_dsql_sqlalchemy.psycopg2:AuroraDSQLDialect_psycopg2"
 # Note: psyscopg (psycopg3) 
-"auroradsql.psycopg" = "auroradsql_sqlalchemy.psycopg:AuroraDSQLDialect_psycopg" 
+"auroradsql.psycopg" = "aurora_dsql_sqlalchemy.psycopg:AuroraDSQLDialect_psycopg" 
 
 [tool.setuptools]
-packages = ["auroradsql_sqlalchemy"]
+packages = ["aurora_dsql_sqlalchemy"]
 
 [tool.pytest.ini_options]
 addopts = "-v  -r chars -s"
