+++ test/test_compiler.py
@@ -13,8 +13,8 @@ from sqlalchemy.testing import fixtures
 from sqlalchemy.testing.assertions import AssertsCompiledSQL
 from sqlalchemy.testing.suite.test_dialect import testing
 
-from auroradsql_sqlalchemy.psycopg import AuroraDSQLDialect_psycopg
-from auroradsql_sqlalchemy.psycopg2 import AuroraDSQLDialect_psycopg2
+from aurora_dsql_sqlalchemy.psycopg import AuroraDSQLDialect_psycopg
+from aurora_dsql_sqlalchemy.psycopg2 import AuroraDSQLDialect_psycopg2
 
 
 class CompileTest(fixtures.TestBase, AssertsCompiledSQL):
